Eliminating MDNode distinction for LoopID.
Since LoopID is no longer a loop identifier, this change allows loop
metadata to be uniqued and also removes the requirement that the first
element of LoopID be self-referential (originally required before distinct
was available).  LoopID is also renamed to LoopMD to reflect that the
variable is loop metadata, not an identifier.